requirements.txt
pyzmq>=25.0.0
rich